Fix payflow template paths and clean core tags
