+++ templates/components/payflow/pagination.html
@@ -1,4 +1,4 @@
-{% load core_tags %}
+{% load payflow_tags %}
 {% if page_obj.has_other_pages %}
 <div id="pagination" class="flex flex-row justify-center items-center w-full gap-x-2">
     {% if page_obj.has_previous %}
