+++ templates/components/payflow/ticket_container.html
@@ -1,4 +1,4 @@
-{% load core_tags %}
+{% load payflow_tags %}
 
 
 {% if expanded %}
