+++ templates/components/payflow/ticket_empty.html
