+++ templates/components/payflow/ticket_message.html
