+++ templates/components/payflow/ticket_status.html
@@ -1,4 +1,4 @@
-{% load core_tags %}
+{% load payflow_tags %}
 
 {% comment %}
 Ticket Status - Muestra el estado actual y la acción esperada
