+++ templates/welp_payflow/attachment.html
@@ -1,5 +1,4 @@
 {% extends "core/base.html" %}
-{% load core_tags %}
 {% load format_tags %}
 
 {% block title %}{{ attachment.file.name|basename }} - Welp Payflow{% endblock %}
