+++ templates/welp_payflow/create.html
@@ -1,5 +1,5 @@
 {% extends "core/base.html" %}
-{% load core_tags %}
+{% load payflow_tags %}
 {% load django_vite %}
 
 {% block title %}Crear Solicitud - Welp Payflow{% endblock %}
