+++ templates/welp_payflow/detail.html
@@ -1,5 +1,5 @@
 {% extends "core/base.html" %}
-{% load core_tags %}
+{% load payflow_tags %}
 {% load django_vite %}
 
 {% block title %}{{ ticket.title }} - Welp Payflow{% endblock %}
