+++ templates/welp_payflow/list.html
@@ -1,5 +1,5 @@
 {% extends "core/base.html" %}
-{% load core_tags %}
+{% load payflow_tags %}
 {% load django_vite %}
 
 {% block title %}Solicitudes - Welp Payflow{% endblock %}
