+++ templates/welp_payflow/success.html
@@ -1,5 +1,5 @@
 {% extends "core/base.html" %}
-{% load core_tags %}
+{% load payflow_tags %}
 {% load django_vite %}
 
 {% block title %}Solicitud Creada - Welp Payflow{% endblock %}
